requirements.txt
flask==2.3.3
transformers==4.35.2
torch==2.1.1
spotipy==2.23.0
python-dotenv==1.0.0 